# Improve writing single characters with UTF8 writer